reorganized tests, added subsequents requests
