+++ apps/media/tests/test_artists.py
